Move create_aggregation into from aggregate.rs to tendermint.rs
Also make it take the id of the to be started aggregation instead of only allowing the current one.
Remove comment from a previous update.
